requirements.txt
python-dotenv
streamlit
openai
PyPDF2
anthropic
python-docx